requirements.txt
torch torchvision --index-url https://download.pytorch.org/whl/cu128
safetensors
compel
pillow
transformers
diffusers
accelerate